Split back into two files
